# src/docs/openapi.yaml
openapi: 3.0.3
info:
  title: 库存调拨规划器 API
  version: 1.0.0
  description: |
    库存调拨规划器提供完整的调拨申请、库存约束检查、调拨方案生成与确认、多仓自动分配、
    仓库管理、SKU 管理、库存管理及调拨记录查询等能力。

    **主要模块：**
    - 仓库管理 (`/api/warehouses`)
    - SKU 管理 (`/api/skus`)
    - 库存管理 (`/api/inventory`)
    - 调拨申请 (`/api/transfer-requests`)
    - 调拨方案 (`/api/transfer-plans`)
    - 调拨记录 (`/api/transfer-records`)

    **约束分级：**
    - `FATAL`：不可通过，阻断调拨
    - `WARNING`：可确认但需业务复核
    - `INFO`：提示性信息

    **调拨方案状态机：**
    - `PENDING`（待确认）→ `CONFIRMED`（已确认）→ `COMPLETED`（已完成）
    - `PENDING` → `REJECTED`（已驳回）
    - 任意状态 → `CANCELLED`（已取消）
  contact:
    name: Inventory Ops Team
servers:
  - url: http://localhost:3000
    description: 本地开发环境
  - url: /
    description: 当前部署
...
paths:
...
  /api/transfer-requests:
    get:
      tags: [Transfer Requests]
      summary: 查询调拨申请列表
      operationId: listTransferRequests
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [DRAFT, SUBMITTED, PLANNING, APPROVED, PROCESSING, COMPLETED, CANCELLED, REJECTED]
        - name: priority
          in: query
          schema:
            type: string
            enum: [LOW, NORMAL, HIGH, URGENT]
        - name: source_warehouse_id
          in: query
          schema: { type: integer, minimum: 1 }
        - name: target_warehouse_id
          in: query
          schema: { type: integer, minimum: 1 }
        - name: keyword
          in: query
          schema: { type: string, maxLength: 50 }
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize50'
      responses:
        '200': { description: 分页申请列表 }
    post:
      tags: [Transfer Requests]
      summary: 创建调拨申请
      operationId: createTransferRequest
      description: 支持创建后直接提交（auto_submit=true），不指定source_warehouse_id触发多仓分配
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/TransferRequestCreate' }
      responses:
        '201': { description: 创建成功 }
        '400': { $ref: '#/components/responses/ValidationError' }
...
  /api/transfer-requests/{id}:
    parameters:
      - $ref: '#/components/parameters/IdParam'
    get:
      tags: [Transfer Requests]
      summary: 查询申请详情（含明细 + 关联方案）
      operationId: getTransferRequest
      responses:
        '200': { description: 详情 }
        '404': { $ref: '#/components/responses/NotFound' }
    put:
      tags: [Transfer Requests]
      summary: 更新调拨申请
      operationId: updateTransferRequest
      description: 仅 DRAFT/SUBMITTED/PLANNING/REJECTED/CANCELLED 状态可编辑
      responses:
        '200': { description: 更新成功 }
    delete:
      tags: [Transfer Requests]
      summary: 删除调拨申请
      operationId: deleteTransferRequest
      description: 仅 DRAFT/CANCELLED/REJECTED 且无关联方案时可删除
      responses:
        '200': { description: 删除成功 }
...
components:
  parameters:
    IdParam:
      name: id
      in: path
      required: true
      schema: { type: integer, minimum: 1, example: 1 }
    Page:
      name: page
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
...
    PageSize50:
      name: pageSize
      in: query
      schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
...
  responses:
    ValidationError:
      description: 请求参数校验失败
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationErrorBody'
    NotFound:
      description: 资源不存在
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorBody'
...
  schemas:
...
    TransferRequestItem:
      type: object
      required: [sku_id, requested_qty]
      properties:
        sku_id: { type: integer, minimum: 1 }
        requested_qty: { type: integer, minimum: 1 }
        unit_price: { type: number, minimum: 0 }
        remark: { type: string, maxLength: 200 }
    TransferRequestCreate:
      type: object
      required: [target_warehouse_id, items]
      properties:
        source_warehouse_id: { type: integer, minimum: 1, description: 不指定则触发多仓自动分配 }
        target_warehouse_id: { type: integer, minimum: 1 }
        request_type: { type: string, enum: [NORMAL, URGENT, EMERGENCY], default: NORMAL }
        priority: { type: string, enum: [LOW, NORMAL, HIGH, URGENT], default: NORMAL }
        requester: { type: string, maxLength: 50 }
        department: { type: string, maxLength: 50 }
        reason: { type: string, maxLength: 500 }
        expected_date: { type: string, format: date }
        items:
          type: array
          minItems: 1
          items: { $ref: '#/components/schemas/TransferRequestItem' }
        auto_submit: { type: boolean, default: false, description: 创建即提交 }
...
    ErrorBody:
      type: object
      required: [error, code]
      properties:
        error: { type: string, description: 面向用户的错误消息 }
        code:
          type: string
          description: 稳定的机器可识别错误码
          example: WAREHOUSE_NOT_FOUND
        requestId:
          type: string
          description: 请求追踪 ID，用于日志关联
        timestamp: { type: string, format: date-time }
    ValidationErrorBody:
      allOf:
        - $ref: '#/components/schemas/ErrorBody'
        - type: object
          properties:
            errors:
              type: array
              items:
                type: object
                required: [path, message]
                properties:
                  path: { type: string, example: target_warehouse_id }
                  message: { type: string, example: target_warehouse_id 必填 }
                  value: { type: any }